separate test for packages
