+++ src/test/java/executor/service/model/proxy/ProxyCredentialsTest.java
@@ -1,4 +1,4 @@
-package executor.service.model;
+package executor.service.model.proxy;
 
 import executor.service.model.proxy.ProxyCredentials;
 import org.junit.jupiter.api.AfterEach;
