+++ src/test/java/executor/service/model/proxy/ProxyNetworkConfigTest.java
@@ -1,4 +1,4 @@
-package executor.service.model;
+package executor.service.model.proxy;
 
 import executor.service.model.proxy.ProxyNetworkConfig;
 import org.junit.jupiter.api.AfterEach;
