+++ src/test/java/executor/service/model/scenario/StepTest.java
@@ -1,4 +1,4 @@
-package executor.service.model;
+package executor.service.model.scenario;
 import executor.service.model.scenario.Step;
 import org.junit.jupiter.api.Test;
